Solution to Find Student with Minimum Marks

+++ Find Student with Minimum Marks@@67877d61dc107486ad5c2749/code.c
@@ -9,12 +9,12 @@ struct info{
 
 int main() {
     int n;
-    float minMarks=0;
+    int minMarks=0;
     scanf("%d",&n);
     struct info e[n];
     for(int i=0;i<n;i++){
         scanf("%d %s %f",&e[i].rollno,e[i].name,&e[i].marks);
-        
+
     if(e[i].marks < e[minMarks].marks){
         minMarks=i;
     }
